// packages/programs/src/splitter/generated/types/MerkleTree.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type MerkleTree = {
  levels: number
  nextIndex: beet.bignum
  root: number[] /* size: 32 */
  filledSubtrees: number[] /* size: 32 */[]
  zeroValues: number[] /* size: 32 */[]
}

/**
 * @category userTypes
 * @category generated
 */
export const merkleTreeBeet = new beet.FixableBeetArgsStruct<MerkleTree>(
  [
    ['levels', beet.u8],
    ['nextIndex', beet.u64],
    ['root', beet.uniformFixedSizeArray(beet.u8, 32)],
    ['filledSubtrees', beet.array(beet.uniformFixedSizeArray(beet.u8, 32))],
    ['zeroValues', beet.array(beet.uniformFixedSizeArray(beet.u8, 32))],
  ],
  'MerkleTree'
)